Add files via upload
Repository Title: ML-projects

Description:
Welcome to ML-projects! This repository holds all of my ML projects . This project hosts a Multi-Linear Regression model that takes into account a student's CAT-1 (out of 35), CAT-2 (out of 35) marks, and Sessional Marks (out of 70) to predict their university marks (out of 70). To simplify the model, CAT-1 and CAT-2 marks are combined into Total CAT Marks (out of 70). With remarkable accuracy and a foundation built on Python's Pandas, Numpy, and Faker libraries, this model is a valuable tool for both students and educators. Explore the code, contribute, and let's work together to enhance academic outcomes!